- Updated required towny version to 0.97.0.23 - Removed the special listener for 0.97.0.14   - Since the required towny version is now higher than that,     it is no longer necessary to be separate from the nation event listener
