+++ src/main/java/com/gmail/goosius/siegewar/listeners/SiegeWarNationEventListener.java
@@ -22,12 +22,7 @@ import com.palmergames.bukkit.towny.event.NationBonusCalculationEvent;
 import com.palmergames.bukkit.towny.event.NationPreRemoveEnemyEvent;
 import com.palmergames.bukkit.towny.event.PreDeleteNationEvent;
 import com.palmergames.bukkit.towny.event.NationPreAddTownEvent;
-import com.palmergames.bukkit.towny.event.nation.NationRankAddEvent;
-import com.palmergames.bukkit.towny.event.nation.NationPreTownLeaveEvent;
-import com.palmergames.bukkit.towny.event.nation.NationListDisplayedNumOnlinePlayersCalculationEvent;
-import com.palmergames.bukkit.towny.event.nation.NationListDisplayedNumTownsCalculationEvent;
-import com.palmergames.bukkit.towny.event.nation.NationListDisplayedNumResidentsCalculationEvent;
-import com.palmergames.bukkit.towny.event.nation.NationListDisplayedNumTownBlocksCalculationEvent;
+import com.palmergames.bukkit.towny.event.nation.*;
 import com.palmergames.bukkit.towny.event.statusscreen.NationStatusScreenEvent;
 import com.palmergames.bukkit.towny.event.townblockstatus.NationZoneTownBlockStatusEvent;
 import com.palmergames.bukkit.towny.object.Nation;
@@ -42,6 +37,7 @@ import org.bukkit.event.Listener;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Comparator;
 import java.util.List;
 
 /**
@@ -53,7 +49,27 @@ public class SiegeWarNationEventListener implements Listener {
 
 	@SuppressWarnings("unused")
 	private final SiegeWar plugin;
-	
+
+	private static final Comparator<Nation> BY_NUM_RESIDENTS = (n1, n2) -> {
+        return SiegeWarNationUtil.getEffectiveNation(n2).getResidents().size() 
+          	   - SiegeWarNationUtil.getEffectiveNation(n1).getResidents().size();
+    };
+
+    private static final Comparator<Nation> BY_TOWNS= (n1, n2) -> {
+        return SiegeWarNationUtil.getEffectiveNation(n2).getNumTowns() 
+          	   - SiegeWarNationUtil.getEffectiveNation(n1).getNumTowns();
+    };
+
+    private static final Comparator<Nation> BY_TOWNBLOCKS= (n1, n2) -> {
+        return SiegeWarNationUtil.getEffectiveNation(n2).getNumTownblocks() 
+          	   - SiegeWarNationUtil.getEffectiveNation(n1).getNumTownblocks();
+    };
+
+    private static final Comparator<Nation> BY_ONLINE= (n1, n2) -> {    
+		return TownyAPI.getInstance().getOnlinePlayers(SiegeWarNationUtil.getEffectiveNation(n2)).size() 
+			   - TownyAPI.getInstance().getOnlinePlayers(SiegeWarNationUtil.getEffectiveNation(n1)).size();
+    };
+
 	public SiegeWarNationEventListener(SiegeWar instance) {
 
 		plugin = instance;
@@ -330,5 +346,39 @@ public class SiegeWarNationEventListener implements Listener {
 			event.setCancelled(true);
 			event.setCancelMessage(Translation.of("plugin_prefix") + Translation.of("msg_err_siege_affected_home_nation_cannot_recruit"));
 		}
+	}
+
+	/**
+	 * Re-Sorts the nations list when Towny sorts it
+	 * 
+	 * - Towny uses this list for the /n list display
+	 * 
+	 * - SiegeWar re-orders it to account for town occupation
+	 *   - Unoccupied towns are counted as part of their natural nation
+	 *   - Occupied towns are counted as part of the occupying nation
+	 */
+	@EventHandler
+	public void on(DisplayedNationsListSortEvent event) {
+		//Get originally sorted list
+		List<Nation> nationList = event.getNations();
+		//Re-sort list, taking occupation into account
+		switch (event.getComparatorType()) {
+			case RESIDENTS:
+				nationList.sort(BY_NUM_RESIDENTS);
+				break;
+			case TOWNBLOCKS:
+				nationList.sort(BY_TOWNBLOCKS);
+				break;
+			case ONLINE:
+				nationList.sort(BY_ONLINE);
+				break;
+			case TOWNS:
+				nationList.sort(BY_TOWNS);
+				break;
+			default:
+				return;	
+		}
+		//Give the re-sorted list to the event object
+		event.setNations(nationList);
 	}
 }
